// src/lib/database.types.ts
export type Database = {
	public: {
		Tables: {
			notification_log: {
				Row: {
					created_at: string;
					delivery_method: Database["public"]["Enums"]["delivery_method"];
					error: string | null;
					error_code: string | null;
					id: string;
					message: string | null;
					message_delivered: boolean;
					type: string;
					updated_at: string;
					user_id: string;
				};
// ...
			};
// ...
			users: {
				Row: {
					created_at: string;
					daily_digest_enabled: boolean;
					daily_digest_notification_time: number;
					email: string;
					email_notifications_enabled: boolean;
					full_phone: string | null;
					id: string;
					next_send_at: string | null;
					phone_country_code: string | null;
					phone_number: string | null;
					phone_verified: boolean;
					sms_notifications_enabled: boolean;
					sms_opted_out: boolean;
					timezone: string;
					updated_at: string;
				};
				Insert: {
					created_at?: string;
					daily_digest_enabled?: boolean;
					daily_digest_notification_time?: number;
					email: string;
					email_notifications_enabled?: boolean;
					full_phone?: string | null;
					id?: string;
					next_send_at?: string | null;
					phone_country_code?: string | null;
					phone_number?: string | null;
					phone_verified?: boolean;
					sms_notifications_enabled?: boolean;
					sms_opted_out?: boolean;
					timezone?: string;
					updated_at?: string;
				};
				Update: {
					created_at?: string;
					daily_digest_enabled?: boolean;
					daily_digest_notification_time?: number;
					email?: string;
					email_notifications_enabled?: boolean;
					full_phone?: string | null;
					id?: string;
					next_send_at?: string | null;
					phone_country_code?: string | null;
					phone_number?: string | null;
					phone_verified?: boolean;
					sms_notifications_enabled?: boolean;
					sms_opted_out?: boolean;
					timezone?: string;
					updated_at?: string;
				};
				Relationships: [
					{
						foreignKeyName: "users_timezone_fkey";
						columns: ["timezone"];
						isOneToOne: false;
						referencedRelation: "timezones";
						referencedColumns: ["value"];
					},
				];
			};
		};
// ...
	};
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
	keyof Database,
	"public"
>];
// ...
export type Enums<
	DefaultSchemaEnumNameOrOptions extends
		| keyof DefaultSchema["Enums"]
		| { schema: keyof DatabaseWithoutInternals },
	EnumName extends DefaultSchemaEnumNameOrOptions extends {
		schema: keyof DatabaseWithoutInternals;
	}
		? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
		: never = never,
> = DefaultSchemaEnumNameOrOptions extends {
	schema: keyof DatabaseWithoutInternals;
}
	? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
	: DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
		? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
		: never;